// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  SELLER: 'SELLER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const OrderStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED',
  PARTIALLY_REFUNDED: 'PARTIALLY_REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const PaymentMethod = {
  STRIPE: 'STRIPE',
  PAYPAL: 'PAYPAL',
  COD: 'COD',
  BANK_TRANSFER: 'BANK_TRANSFER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const ShippingStatus = {
  NOT_SHIPPED: 'NOT_SHIPPED',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
  RETURNED: 'RETURNED'
} as const

export type ShippingStatus = (typeof ShippingStatus)[keyof typeof ShippingStatus]


export const NotificationType = {
  ORDER_UPDATE: 'ORDER_UPDATE',
  PAYMENT: 'PAYMENT',
  PROMOTION: 'PROMOTION',
  SYSTEM: 'SYSTEM',
  REVIEW: 'REVIEW',
  WISHLIST: 'WISHLIST'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const CouponType = {
  PERCENTAGE: 'PERCENTAGE',
  FIXED_AMOUNT: 'FIXED_AMOUNT',
  FREE_SHIPPING: 'FREE_SHIPPING'
} as const

export type CouponType = (typeof CouponType)[keyof typeof CouponType]


export const BannerPosition = {
  HERO: 'HERO',
  SIDEBAR: 'SIDEBAR',
  POPUP: 'POPUP',
  MIDDLE: 'MIDDLE'
} as const

export type BannerPosition = (typeof BannerPosition)[keyof typeof BannerPosition]


export const BlogStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type BlogStatus = (typeof BlogStatus)[keyof typeof BlogStatus]


export const ContactStatus = {
  UNREAD: 'UNREAD',
  READ: 'READ',
  REPLIED: 'REPLIED',
  RESOLVED: 'RESOLVED'
} as const

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus]